Added 'Winston' module for logging
Added the Winston node.js module and setup simple logging to file.  'npm
install winston' to include. This first instance will log to a file
named standardLog.log in JSON format for a size up to 10MB in a folder
named 'logs'. Once the data cap is hit the file will begin to roll.
https://github.com/winstonjs/winston

https://github.com/winstonjs/winston/blob/master/docs/transports.md#file-transport