Minor clean up
Removed the fs, path, and process modules from main file as they were no
longer being used. They were moved into the cena module. Moved the
RPSManager down - this time the bot is able to initialize before it is
called so it shouldn't be a problem. Also removed a commented out line
from the tableCatcher as it was not needed.