Changed name of cenaModule.js to memeRepliesModule.js
Changed name of the file as the cenaModule is going to become the file
that handles all shitposts. Any code snippet that simply listens to
messages and replies to a user with text and or images will be housed
here. Changed relevant lines in sdgdiscordbot.js so that the cena
replies would still work.