MongoDB and Karma System
Got caught up in hashing this code out and didn't commit throughout the
process. This code introduces MongoDB into shamebot. However, most of
the code is used for parsing through messages to find mentions followed
by either '++' or '--' for upvoting and downvoting users on the server.
This will serve as the basis for our karma system. The upvotes and
downvotes, when found, increment and decrement a document(record) in the
database for the user on the server. Karma is server specific so that a
user can have different karma level depending on the server they are on.
The parsing system allows for natural messaging and up/downvotes do not
need to be issued as a command. The votes can be given anywhere in a
message with other text as long as they are followed by a ++ or --.
Additionally, votes can be given to multiple users in the same message
in as long as the follow the [user] [++/--] convention. Each vote will
open a new connection to the DB as I was not able to get a connection
pool working. That will implemented in a future push. Finally, there is
a new command: !honor [mention]. This will pull up the net karma for the
mentioned user.
Next steps: Modularize this new code and get connection pools going.
Also need to put in some additional checks so that the bot will not
crash if the DB is not available.